micro: port operator BATCH_MATMUL kernel from lite with test
Complete implementation of TFLM operator BATCH_MATMUL and associated TFLM test code.

PR step 5 of the work to port operator BATCH_MATMUL as tracked in Issue #46504
